Correcion EmpleadoDAO / modificacion en EmpleadoService / cambio de tipo de dato en id (de int a long)

+++ TPIProgramacionII/src/tpiprogramacionii/dao/GenericDAO.java
@@ -8,13 +8,13 @@ public interface GenericDAO<T> {
     // Métodos autónomos (crean su propia conexión)
     void insertar(T entidad) throws Exception;
     void actualizar(T entidad) throws Exception;
-    void eliminar(int id) throws Exception;
-    T leer(int id) throws Exception;
+    void eliminar(Long id) throws Exception;
+    T leer(Long id) throws Exception;
     List<T> leerTodos() throws Exception;
     
     // Métodos transaccionales (reciben conexión externa)
     void insertTx(T entidad, Connection conn) throws Exception;
     void actualizarTx(T entidad, Connection conn) throws Exception;
-    void eliminarTx(int id, Connection conn) throws Exception;
+    void eliminarTx(Long id, Connection conn) throws Exception;
 
 }
